moved data specific functions to their own files, snet and consumption processors and routers updated
