+++ db/index.ts
@@ -2,6 +2,9 @@
 export * from './connect_db';
 export * from './authenticate';
 export * from './queue';
+export * from './supplemental';
+export * from './snet';
+export * from './consumption';
 
 export const supplementalTables: string[] = [
     'sciname', 'countries', 'production', 'products', 'baci'
